Fix windows build (size_t vs. int) some more.
TBR=mstarzinger@chromium.org
BUG=

Review URL: https://codereview.chromium.org/336803002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@21845 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/ast-value-factory.cc
@@ -77,7 +77,7 @@ bool AstString::AsArrayIndex(uint32_t* index) const {
 
 
 bool AstString::IsOneByteEqualTo(const char* data) const {
-  int length = strlen(data);
+  int length = static_cast<int>(strlen(data));
   if (is_one_byte_ && literal_bytes_.length() == length) {
     const char* token = reinterpret_cast<const char*>(literal_bytes_.start());
     return !strncmp(token, data, length);
